requirements.txt
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-httplib2>=0.1.0
google-cloud-resource-manager>=1.0.0
google-cloud-iam>=2.0.0
google-cloud-asset>=3.0.0
google-cloud-logging>=3.0.0
networkx>=2.6
pyyaml>=5.4
click>=8.0
rich>=10.0
tqdm>=4.60
pyvis>=0.2.0
matplotlib>=3.4
pandas>=1.3
numpy>=1.21 